ffi alignment

+++ operator/halo2kzg/lib/halo2kzg.h
@@ -1,8 +1,8 @@
 #include <stdbool.h>
-#include <stdint.h>
+#include <stddef.h>
 
-bool verify_halo2_kzg_proof_ffi(unsigned char *proof_bytes, uint64_t proof_len,
-                                unsigned char *constraint_system_bytes, uint64_t cs_len,
-                                unsigned char *verifier_params_bytes, uint64_t vk_len,
-                                unsigned char *kzg_params_bytes, uint64_t kzg_params_len,
-                                unsigned char *public_input_bytes, uint64_t public_input_len);
+bool verify_halo2_kzg_proof_ffi(unsigned char *proof_bytes, size_t proof_len,
+                                unsigned char *constraint_system_bytes, size_t cs_len,
+                                unsigned char *verifier_params_bytes, size_t vk_len,
+                                unsigned char *kzg_params_bytes, size_t kzg_params_len,
+                                unsigned char *public_input_bytes, size_t public_input_len);
